+++ config/routes.rb
@@ -1,6 +1,4 @@
 RailsUrlShortener::Engine.routes.draw do
-  constraints RailsUrlShortener::HostConstraint do
-    get '/', to: RailsUrlShortener::RootHandler, constraints: ->(req) { RailsUrlShortener.block_root }
-    get '/:key', to: 'urls#show'
-  end
+  get '/', to: RailsUrlShortener::RootHandler if RailsUrlShortener.block_root
+  get '/:key', to: 'urls#show'
 end